requirements.txt
Flask==3.0.0
Flask-Cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
waitress==2.1.2